add feedback to the user when bad credentials (#16)
One tip here - if you'll merge something in the future, don't forget to squash commits before merging. This action simply transform all commits into the single one. Later on it's better for git history etc. You'll likely want to read 1 commit message instead of 15 describing basically the same as the mentioned 1 commit. Not sure if you're familiar with it so just leaving there a note. :)